restructured modem part
